requirements.txt
# requirements for mynas<=0.4.1

Click==7.0
graphviz==0.10.1
imageio==2.5.0
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
psutil==5.6.2
pytest==4.5.0
# pytest==3.9.1 # strange... Python 3.7.0 sometimes donot work with pytest==4.5.0; can use 3.9.1
pytest-cov==2.7.1
PyYAML==5.1
scipy==1.3.0rc1 # Python >= 3.5 required
setproctitle==1.1.10
six==1.12.0
tensorboardX==1.6
torch==1.2.0
torchvision==0.4.0

-e .[vis,det]
